api/requirements.txt
langchain
cchardet
chardet
openai
tiktoken
chromadb
pinecone-client
unstructured
pypdf
python-multipart